server-ELD/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
geopy==2.4.1
python-dateutil==2.8.2
reportlab
